Refactor models to reduce duplication
To be able to have `models/lib.rs` with shared code, we needed to stop symlinking the entire `src` directory and instead treat the variant-specific code as Rust modules, symlinking the correct one to `variant::current` so that `lib.rs` can re-export `Settings`.  (We also have to symlink in `variant/mod.rs` so Rust knows it's part of the module hierarchy.)

Now, each variant defines a `Settings` structure using common sub-structures.  Those sub-structures live in the main `lib.rs`, and the modeled types are now in a shared `modeled_types` module.  No changes were made to the code, just its organization.

Documentation was updated to reflect these changes, including using cargo-readme to generate README.md now that there's a single source.
