+++ workspaces/models/.gitignore
@@ -1,1 +1,2 @@
-/current/src
+/src/variant/current
+/src/variant/mod.rs
