+++ workspaces/models/Cargo.toml
@@ -3,6 +3,8 @@ name = "models"
 version = "0.1.0"
 authors = ["Tom Kirchner <tjk@amazon.com>"]
 edition = "2018"
+publish = false
+build = "build.rs"
 
 [dependencies]
 base64 = "0.10"
@@ -13,6 +15,11 @@ snafu = "0.5"
 toml = "0.5"
 url = "2.1"
 
+[build-dependencies]
+cargo-readme = "3.1"
+
 [lib]
+# We're picking the current *model* with build.rs, so users shouldn't think
+# about importing *models* (plural), just the one current model.
 name = "model"
-path = "current/src/model.rs"
+path = "src/lib.rs"
