+++ workspaces/models/src/modeled_types/shared.rs
@@ -1,123 +1,54 @@
-//! This module contains data types that can be used in the model when special input/output
-//! (ser/de) behavior is desired.  For example, the SingleLineString type can be used for a model field
-//! when we don't even want to accept an API call with multiple lines in the input.
-
-// The pattern in this file is to make a struct and implement TryFrom<&str> with code that does
-// necessary checks and returns the struct.  Other traits that treat the struct like a string can
-// be implemented for you with the string_impls_for macro.
-
 use serde::{Deserialize, Deserializer, Serialize, Serializer};
 // Just need serde's Error in scope to get its trait methods
 use serde::de::Error as _;
-use snafu::ensure;
+use snafu::{ensure, ResultExt};
 use std::borrow::Borrow;
 use std::convert::TryFrom;
 use std::fmt;
 use std::ops::Deref;
+use super::error;
 
-pub mod error {
-    use snafu::Snafu;
+/// ValidBase64 can only be created by deserializing from valid base64 text.  It stores the
+/// original text, not the decoded form.  Its purpose is input validation, namely being used as a
+/// field in a model structure so that you don't even accept a request with a field that has
+/// invalid base64.
+// Note: we use the default base64::STANDARD config which uses/allows "=" padding.
+#[derive(Debug, Clone, Eq, PartialEq, Hash)]
+pub struct ValidBase64 {
+    inner: String,
+}
 
-    #[derive(Debug, Snafu)]
-    #[snafu(visibility = "pub(super)")]
-    pub enum Error {
-        #[snafu(display("Can't create SingleLineString containing line terminator"))]
-        StringContainsLineTerminator,
+/// Validate base64 format before we accept the input.
+impl TryFrom<&str> for ValidBase64 {
+    type Error = error::Error;
 
-        #[snafu(display("Invalid base64 input: {}", source))]
-        InvalidBase64 { source: base64::DecodeError },
-
-        #[snafu(display(
-            "Identifiers may only contain ASCII alphanumerics plus hyphens, received '{}'",
-            input
-        ))]
-        InvalidIdentifier { input: String },
-
-        #[snafu(display("Given invalid URL '{}'", input))]
-        InvalidUrl { input: String },
-
-        // Some regexes are too big to usefully display in an error.
-        #[snafu(display("{} given invalid input: {}", thing, input))]
-        BigPattern { thing: String, input: String },
-
-        #[snafu(display("Given invalid cluster name '{}': {}", name, msg))]
-        InvalidClusterName { name: String, msg: String },
+    fn try_from(input: &str) -> Result<Self, Self::Error> {
+        base64::decode(&input).context(error::InvalidBase64)?;
+        Ok(ValidBase64 {
+            inner: input.to_string(),
+        })
     }
 }
 
-/// Helper macro for implementing the common string-like traits for a modeled type.
-/// Pass the name of the type, and the name of the type in quotes (to be used in string error
-/// messages, etc.).
-macro_rules! string_impls_for {
-    ($for:ident, $for_str:expr) => {
-        impl TryFrom<String> for $for {
-            type Error = error::Error;
+string_impls_for!(ValidBase64, "ValidBase64");
 
-            fn try_from(input: String) -> Result<Self, Self::Error> {
-                Self::try_from(input.as_ref())
-            }
-        }
+#[cfg(test)]
+mod test_valid_base64 {
+    use super::ValidBase64;
+    use std::convert::TryFrom;
 
-        impl<'de> Deserialize<'de> for $for {
-            fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-            where
-                D: Deserializer<'de>,
-            {
-                let original = String::deserialize(deserializer)?;
-                Self::try_from(original).map_err(|e| {
-                    D::Error::custom(format!("Unable to deserialize into {}: {}", $for_str, e))
-                })
-            }
-        }
+    #[test]
+    fn valid_base64() {
+        let v = ValidBase64::try_from("aGk=").unwrap();
+        let decoded_bytes = base64::decode(v.as_ref()).unwrap();
+        let decoded = std::str::from_utf8(&decoded_bytes).unwrap();
+        assert_eq!(decoded, "hi");
+    }
 
-        /// We want to serialize the original string back out, not our structure, which is just there to
-        /// force validation.
-        impl Serialize for $for {
-            fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-            where
-                S: Serializer,
-            {
-                serializer.serialize_str(&self.inner)
-            }
-        }
-
-        impl Deref for $for {
-            type Target = str;
-            fn deref(&self) -> &Self::Target {
-                &self.inner
-            }
-        }
-
-        impl Borrow<String> for $for {
-            fn borrow(&self) -> &String {
-                &self.inner
-            }
-        }
-
-        impl Borrow<str> for $for {
-            fn borrow(&self) -> &str {
-                &self.inner
-            }
-        }
-
-        impl AsRef<str> for $for {
-            fn as_ref(&self) -> &str {
-                &self.inner
-            }
-        }
-
-        impl fmt::Display for $for {
-            fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-                write!(f, "{}", self.inner)
-            }
-        }
-
-        impl From<$for> for String {
-            fn from(x: $for) -> Self {
-                x.inner
-            }
-        }
-    };
+    #[test]
+    fn invalid_base64() {
+        assert!(ValidBase64::try_from("invalid base64").is_err());
+    }
 }
 
 // =^..^=   =^..^=   =^..^=   =^..^=   =^..^=   =^..^=   =^..^=   =^..^=   =^..^=
